refactor(authup): group the dev machinery by what each file is
`src/dev` was twelve flat files whose names disclosed the subject but not
the kind. `rate-limit.ts` reads as the rate limiter when it is the skip
predicate feeding server-core's, and nothing said `open-in-editor.ts` is a
routup handler.

Two folders now say it, mirroring server-core's own
`adapters/http/middleware/built-in`: `server/` for the vite dev server one
per console, `middleware/` for what dev adds to or changes in the request
pipeline. `source.ts` and `resolve.ts` become `package.ts`, since all three
functions answer where a console package is and whether it is source.
