+++ apps/authup/src/dev/server/index.ts
@@ -5,10 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './constants.ts';
-export * from './env.ts';
-export * from './middleware/index.ts';
-export * from './mount.ts';
-export * from './package.ts';
-export * from './server/index.ts';
+export * from './auth.ts';
+export * from './module.ts';
+export * from './static.ts';
 export * from './types.ts';
